Drop platform versions [SDK-2895] (#527)
* Bump platform versions

* Remove platform checks

* Remove old Web Auth strategies

* Merge SessionTransaction into BaseAuthTransaction

* Remove ControllerModalPresenter

* Update Podspec

* Update pbxproj

* Rename auth transaction classes

* Add missing import

* Format intializer calls

* Document the changes
