+++ Auth0.podspec
@@ -5,16 +5,16 @@ web_auth_files = [
   'Auth0/ObjectiveC/A0RSA.m',
   'Auth0/ObjectiveC/A0SHA.h',
   'Auth0/ObjectiveC/A0SHA.m',
-  'Auth0/A0SimpleKeychain+RSAPublicKey.swift',
   'Auth0/Array+Encode.swift',
+  'Auth0/ASCallbackTransaction.swift',
+  'Auth0/ASTransaction.swift',
   'Auth0/AuthCancelable.swift',
   'Auth0/AuthProvider.swift',
   'Auth0/AuthSession.swift',
   'Auth0/AuthTransaction.swift',
-  'Auth0/AuthenticationServicesSession.swift',
-  'Auth0/AuthenticationServicesSessionCallback.swift',
-  'Auth0/BaseAuthTransaction.swift',
-  'Auth0/BaseWebAuth.swift',
+  'Auth0/Auth0WebAuth.swift',
+  'Auth0/BaseCallbackTransaction.swift',
+  'Auth0/BaseTransaction.swift',
   'Auth0/BioAuthentication.swift',
   'Auth0/ClaimValidators.swift',
   'Auth0/IDTokenSignatureValidator.swift',
@@ -27,19 +27,14 @@ web_auth_files = [
   'Auth0/NSURLComponents+OAuth2.swift',
   'Auth0/OAuth2Grant.swift',
   'Auth0/ResponseType.swift',
-  'Auth0/SessionCallbackTransaction.swift',
-  'Auth0/SessionTransaction.swift',
   'Auth0/TransactionStore.swift',
-  'Auth0/WebAuthenticatable.swift',
+  'Auth0/WebAuth.swift',
   'Auth0/WebAuthError.swift',
   'Auth0/_ObjectiveWebAuth.swift'
 ]
 
 ios_files = [
-  'Auth0/ControllerModalPresenter.swift',
   'Auth0/MobileWebAuth.swift',
-  'Auth0/SafariSession.swift',
-  'Auth0/SilentSafariViewController.swift',
   'Auth0/UIApplication+Shared.swift'
 ]
 
@@ -56,7 +51,7 @@ excluded_files = [
 
 Pod::Spec.new do |s|
   s.name             = 'Auth0'
-  s.version          = '1.38.0'
+  s.version          = '2.0.0-beta'
   s.summary          = "Swift toolkit for Auth0 API"
   s.description      = <<-DESC
                         Auth0 API toolkit written in Swift for iOS, watchOS, tvOS & macOS apps
@@ -67,10 +62,10 @@ Pod::Spec.new do |s|
   s.source           = { :git => 'https://github.com/auth0/Auth0.swift.git', :tag => s.version.to_s }
   s.social_media_url = 'https://twitter.com/auth0'
 
-  s.ios.deployment_target = '9.0'
-  s.osx.deployment_target = '10.11'
-  s.watchos.deployment_target = '2.0'
-  s.tvos.deployment_target = '9.0'
+  s.ios.deployment_target = '12.0'
+  s.osx.deployment_target = '10.15'
+  s.watchos.deployment_target = '6.2'
+  s.tvos.deployment_target = '12.0'
   s.requires_arc = true
 
   s.ios.source_files = 'Auth0/*.{swift,h,m}', 'Auth0/ObjectiveC/*.{h,m}'
