--- Auth0/SessionCallbackTransaction.swift
@@ -1,46 +1,0 @@
-// SessionCallbackTransaction.swift
-//
-// Copyright (c) 2020 Auth0 (http://auth0.com)
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-#if WEB_AUTH_PLATFORM
-import Foundation
-
-class SessionCallbackTransaction: NSObject, AuthTransaction {
-
-    var state: String?
-    var authSession: AuthSession?
-    let callback: (Bool) -> Void
-
-    init(callback: @escaping (Bool) -> Void) {
-        self.callback = callback
-    }
-
-    func cancel() {
-        self.callback(false)
-    }
-
-    func handleUrl(_ url: URL) -> Bool {
-        self.callback(true)
-        return true
-    }
-
-}
-#endif
