+++ Auth0Tests/BioAuthenticationSpec.swift
@@ -27,7 +27,6 @@ import LocalAuthentication
 
 @testable import Auth0
 
-@available(iOS 10.0, OSX 10.15, *)
 class BioAuthenticationSpec: QuickSpec {
 
     override func spec() {
