+++ Auth0Tests/ClaimValidatorsSpec.swift
@@ -26,7 +26,6 @@ import Nimble
 
 @testable import Auth0
 
-@available(iOS 10.0, macOS 10.12, *)
 class ClaimValidatorsSpec: IDTokenValidatorBaseSpec {
     
     override func spec() {
