+++ Auth0Tests/CryptoExtensions.swift
@@ -28,14 +28,12 @@ import Auth0ObjectiveC
 
 @testable import Auth0
 
-@available(iOS 10.0, macOS 10.12, *)
 extension SecKey {
     func export() -> Data {
         return SecKeyCopyExternalRepresentation(self, nil)! as Data
     }
 }
 
-@available(iOS 10.0, macOS 10.12, *)
 extension JWTAlgorithm {
     func sign(value: Data, key: SecKey = TestKeys.rsaPrivate) -> Data {
         switch self {
