+++ Auth0Tests/Generators.swift
@@ -27,7 +27,6 @@ import JWTDecode
 
 // MARK: - Keys
 
-@available(iOS 10.0, macOS 10.12, *)
 enum TestKeys {
     static let rsaPrivate: SecKey = {
         let query: [String: Any] = [String(kSecAttrKeyType): kSecAttrKeyTypeRSA,
@@ -114,7 +113,6 @@ private func generateJWTPayload(iss: String?,
     return encodeJWTPart(from: bodyDict)
 }
 
-@available(iOS 10.0, macOS 10.12, *)
 func generateJWT(alg: String = JWTAlgorithm.rs256.rawValue,
                  kid: String? = Kid,
                  iss: String? = "https://tokens-test.auth0.com/",
@@ -181,7 +179,6 @@ private func extractData(from bytes: UnsafePointer<UInt8>) -> (UnsafePointer<UIn
     return (pointer, data)
 }
 
-@available(iOS 10.0, macOS 10.12, *)
 func generateRSAJWK(from publicKey: SecKey = TestKeys.rsaPublic, keyId: String = Kid) -> JWK {
     let asn = { (bytes: UnsafePointer<UInt8>) -> JWK? in
         guard bytes.pointee == 0x30 else { return nil } // Checks that this is a SEQUENCE triplet
