+++ Auth0Tests/IDTokenValidatorSpec.swift
@@ -30,7 +30,6 @@ import OHHTTPStubsSwift
 
 @testable import Auth0
 
-@available(iOS 10.0, macOS 10.12, *)
 class IDTokenValidatorSpec: IDTokenValidatorBaseSpec {
 
     override func spec() {
