+++ Auth0Tests/JWKSpec.swift
@@ -27,7 +27,6 @@ import OHHTTPStubs
 
 @testable import Auth0
 
-@available(iOS 10.0, macOS 10.12, *)
 class JWKSpec: QuickSpec {
     
     override func spec() {
