+++ Auth0Tests/JWTAlgorithmSpec.swift
@@ -26,7 +26,6 @@ import Nimble
 
 @testable import Auth0
 
-@available(iOS 10.0, macOS 10.12, *)
 class JWTAlgorithmSpec: QuickSpec {
 
     override func spec() {
