+++ Auth0Tests/NativeAuthSpec.swift
@@ -40,7 +40,7 @@ private let Timeout: DispatchTimeInterval = .seconds(2)
 private let AccessToken = UUID().uuidString.replacingOccurrences(of: "-", with: "")
 private let FacebookToken = UUID().uuidString.replacingOccurrences(of: "-", with: "")
 private let InvalidFacebookToken = UUID().uuidString.replacingOccurrences(of: "-", with: "")
-@available(iOS 10.0, macOS 10.12, *) private let IdToken = generateJWT().string
+private let IdToken = generateJWT().string
 
 class MockNativeAuthTransaction: NativeAuthTransaction {
     var connection: String
@@ -66,19 +66,11 @@ class MockNativeAuthTransaction: NativeAuthTransaction {
         self.delayed = { _ in }
     }
 
-    #if os(iOS)
-    func resume(_ url: URL, options: [A0URLOptionsKey : Any]) -> Bool {
-        self.delayed(self.onNativeAuth())
-        self.delayed = { _ in }
-        return true
-    }
-    #else
     func resume(_ url: URL) -> Bool {
         self.delayed(self.onNativeAuth())
         self.delayed = { _ in }
         return true
     }
-    #endif
 
     /// Test Hooks
     var onNativeAuth: () -> Result<NativeAuthCredentials> = {
@@ -86,7 +78,6 @@ class MockNativeAuthTransaction: NativeAuthTransaction {
     }
 }
 
-@available(iOS 10.0, macOS 10.12, *)
 class NativeAuthSpec: QuickSpec {
 
     override func spec() {
