+++ README.md
@@ -23,7 +23,7 @@ Swift toolkit that lets you communicate efficiently with many of the [Auth0 API]
 
 ## Requirements
 
-- iOS 9+ / macOS 10.11+ / tvOS 9.0+ / watchOS 2.0+
+- iOS 12+ / macOS 10.15+ / tvOS 12.0+ / watchOS 6.2+
 - Xcode 12.x / 13.x
 - Swift 4.x / 5.x
